Organizando arquivos
Separei os arquivos de lógica das rotas em um componente para que o arquivo de rotas fique fácil de entender
